requirements.txt
typer>=0.9.0
rich>=13.0.0
pyyaml>=6.0
jsonschema>=4.0
pydantic>=1.10
